+++ examples/aibs_smartspim_instrument.py
@@ -2,6 +2,8 @@
 
 import datetime
 
+from aind_data_schema_models.organizations import Organization
+
 from aind_data_schema.components.devices import (
     AdditionalImagingDevice,
     Detector,
@@ -13,7 +15,6 @@ from aind_data_schema.components.devices import (
     ScanningStage,
 )
 from aind_data_schema.core.instrument import Com, Instrument
-from aind_data_schema.models.organizations import Organization
 
 inst = Instrument(
     instrument_id="SmartSPIM2-2",