+++ examples/aind_smartspim_instrument.py
@@ -2,9 +2,10 @@
 
 import datetime
 
+from aind_data_schema_models.organizations import Organization
+
 from aind_data_schema.components.devices import Filter, Laser, MotorizedStage, OpticalTable, ScanningStage
 from aind_data_schema.core.instrument import Com, Detector, Instrument, Objective
-from aind_data_schema.models.organizations import Organization
 
 inst = Instrument(
     instrument_id="SmartSPIM2-1",