+++ examples/bergamo_ophys_session.py
@@ -2,6 +2,8 @@
 
 from datetime import datetime, timezone
 
+from aind_data_schema_models.modalities import Modality
+
 from aind_data_schema.components.stimulus import PhotoStimulation, PhotoStimulationGroup
 from aind_data_schema.core.session import (
     DetectorConfig,
@@ -12,7 +14,6 @@ from aind_data_schema.core.session import (
     StimulusModality,
     Stream,
 )
-from aind_data_schema.models.modalities import Modality
 
 # If a timezone isn't specified, the timezone of the computer running this
 # script will be used as default