+++ examples/mri_session.py
@@ -2,10 +2,11 @@
 
 from decimal import Decimal
 
+from aind_data_schema_models.modalities import Modality
+
 from aind_data_schema.components.coordinates import Rotation3dTransform, Scale3dTransform, Translation3dTransform
 from aind_data_schema.components.devices import Scanner
 from aind_data_schema.core.session import MRIScan, MriScanSequence, ScanType, Session, Stream, SubjectPosition
-from aind_data_schema.models.modalities import Modality
 
 scan1 = MRIScan(
     scan_index="1",