+++ tests/test_rig.py
@@ -3,6 +3,8 @@
 import unittest
 from datetime import date, datetime
 
+from aind_data_schema_models.modalities import Modality
+from aind_data_schema_models.organizations import Organization
 from pydantic import ValidationError
 
 from aind_data_schema.components.devices import (
@@ -28,8 +30,6 @@ from aind_data_schema.components.devices import (
     Patch,
 )
 from aind_data_schema.core.rig import Rig
-from aind_data_schema.models.modalities import Modality
-from aind_data_schema.models.organizations import Organization
 
 
 class RigTests(unittest.TestCase):